build: img,svg 추가 및 ThemaProvider예시코드  가독성을 위해 컴포넌트 분리
