+++ .eslintrc.js
@@ -23,5 +23,6 @@ module.exports = {
   rules: {
     "react/react-in-jsx-scope": 0,
     "no-unused-vars": "warn",
+    "react/prop-types": "off",
   },
 };
