--- src/assets/dummy.txt
@@ -1,1 +1,0 @@
-123
